requirements.txt
# UFDR Copilot — Python dependencies
# Install: pip install -r requirements.txt

# ── Core ──────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Embeddings ────────────────────────────────
sentence-transformers==3.0.1
torch>=2.0.0                    # CPU version; for GPU: pip install torch --index-url https://download.pytorch.org/whl/cu118

# ── Vector Search ─────────────────────────────
faiss-cpu==1.8.0                # CPU FAISS; swap for faiss-gpu if you have CUDA
# faiss-gpu==1.7.4              # uncomment for GPU support

# ── Persistent Vector Store ───────────────────
chromadb==0.5.0

# ── API Server ────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── HTTP Client (for retriever_client.py) ─────
requests==2.32.3

# ── Utilities ────────────────────────────────
tqdm==4.66.4
python-dotenv==1.0.1
